pyproject.toml
[tool.poetry]
name = "AEDS"
version = "0.1.0"
description = "Authenticated Encrypted Data Storage"
authors = ["eliaszeus.rangel@gmail.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
PyNaCl = "^1.5"
Flask = "^2.3"
flask-jwt-extended = "^4.5.2"
 
[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"